Fix: Temp table indexes with name starting without '#' do not follow rollback semantics
We use the is_bbf_temp_table flag in ENR metadata for transaction
behaviour. Currently, we only mark this flag as true if the relname
starts with '#', however all indexes on ENR temp tables should be have
this flag set.
We are refactoring the register_ENR function and marking this flag early
rather than in this method

Signed-off-by: Ayush Shah <ayushdsh@amazon.com>