Fix: Rectified dialect check in ENRGetSystableScan function to also check for the conn_type (#632) (#637)
* Fix: Temp table indexes with name starting without '#' do not follow rollback semantics (#626)

We use the is_bbf_temp_table flag in ENR metadata for transaction
behaviour. Currently, we only mark this flag as true if the relname
starts with '#', however all indexes on ENR temp tables should be have
this flag set.
We are refactoring the register_ENR function and marking this flag early
rather than in this method

Signed-off-by: Ayush Shah <ayushdsh@amazon.com>

* Fix: Rectified dialect check in ENRGetSystableScan function to also check for the conn_type (#632)

In ENRGetSystableScan, if the dialect is not TSQL_DIALECT and there is a
scan initiated for a catalog (other than Sequence), we return false i.e.
the tuple being scanned for does not exist in ENR. But this is not
reliable in case this scan is initiated from a PG function since the
dialect then would be PG_DIALECT and the scan in ENR would return false,
even though the actual tuple being searched for is in ENR (in case it's
a T-SQL temp table). This results in a "could not open a relation" error
since the tuple also does not exist in pg_catalog.

Task: BABEL-6099

Signed-off-by: Ayush Shah <ayushdsh@amazon.com>