Obtain correct first character in pattern for ESCAPE in LIKE (#649) (#660)
Whenever there is ESCAPE in LIKE operator and the pattern is either infix or postfix, Babelfish is unable to handle the wildcard properly.

This commit fixes that by ignoring the ESCAPE character in the pattern and finding the correct pattern.
Extension PR: https://github.com/babelfish-for-postgresql/babelfish_extensions/pull/4251

Task: BABEL-6180

Signed-off-by: Shameem Ahmed shmeeh@amazon.com