[BABEL-6276] Band aid fix for row visibility issue in output clause
Community commit "Add missing EPQ recheck for TID Scan" breaks
babelfish's implementation of output clause as it depends upon self
joining the target table for update with itself with ctid in join
condition. As a short term fix we are disabling this community fix for
babelfish connections and revisit this later on.

cr: https://code.amazon.com/reviews/CR-244285229